Move --automatic-fix functionality into FileCacheEntry
Don't bother having multiple redundant ways for plugins to fetch data
structures for file contents.
