+++ src/Phan/PluginV2.php
@@ -93,7 +93,7 @@ use Phan\PluginV2\IssueEmitter;
  *     This is useful to handle functions getAnalyzeFunctionCallClosures did not pick up
  *
  *     (implement \Phan\PluginV2\HandleLazyLoadInternalFunctionCapability)
- * 15. getAutomaticFixers() : array<string,Closure(CodeBase,FileContents,IssueInstance):(?FileEditSet)>
+ * 15. getAutomaticFixers() : array<string,Closure(CodeBase,FileCacheEntry,IssueInstance):(?FileEditSet)>
  *
  *     This method is called to fetch the issue names the plugin can sometimes automatically fix.
  *     Returns a map from issue name to the closure to generate a fix for instances of that issue.
