+++ src/Phan/PluginV2/AutomaticFixCapability.php
@@ -5,7 +5,7 @@ namespace Phan\PluginV2;
 use Closure;
 use Phan\CodeBase;
 use Phan\IssueInstance;
-use Phan\Plugin\Internal\IssueFixingPlugin\FileContents;
+use Phan\Library\FileCacheEntry;
 use Phan\Plugin\Internal\IssueFixingPlugin\FileEditSet;
 
 /**
@@ -18,7 +18,7 @@ interface AutomaticFixCapability
      * This method is called to fetch the issue names the plugin can sometimes automatically fix.
      * Returns a map from issue name to the closure to generate a fix for instances of that issue.
      *
-     * @return array<string,Closure(CodeBase,FileContents,IssueInstance):(?FileEditSet)>
+     * @return array<string,Closure(CodeBase,FileCacheEntry,IssueInstance):(?FileEditSet)>
      */
     public function getAutomaticFixers() : array;
 }
